refactor: 모달 파일 축소화
